Parallel upgrade fix
Fixed the greedy method upgrade to the parallel algorithm by fixing how the seams are removed from the image. Now, each thread handles a certain number of rows (which makes sure threads won't encroach on one another). For each row, the thread first finds the list of pixels that should not be copied (those in seams), sorts them in ascending order, and removes them from the row.